Project v1/requirements.txt
numpy==1.26.4
scipy==1.14.1
matplotlib==3.9.2
PyYAML==6.0.2
jupyter==1.1.1
pytest==8.3.3
joblib==1.4.2
librosa==0.10.2.post1
pandas==2.2.3
scikit-learn==1.6.1
soundfile==0.12.1
tensorflow==2.17.1
tensorflow-hub==0.16.1
opencv-python-headless==4.10.0.84
tqdm==4.67.1
seaborn==0.13.2
datasets[audio]==3.2.0
gdown==5.2.0
ultralytics==8.3.71
